Only show load button interface if agent has not been loaded
